requirements.txt
# Requirements for Sanctions Screening Platform
# Author: Eon (Himanshu Shekhar)
# Email: eonhimanshu@gmail.com
# GitHub: https://github.com/eonn/sanctions-screening
# Created: 2024

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# NLP and ML
transformers==4.35.2
torch==2.5.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Fuzzy logic and string matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
rapidfuzz==3.5.2

# Database and caching
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
psycopg2-binary==2.9.9

# API and web
requests==2.31.0
httpx==0.25.2
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Security and validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development tools
black==23.11.0
flake8==6.1.0

# Message Queue and Streaming
kafka-python==2.0.2
aiokafka==0.8.1
pika==1.3.2
aio-pika==9.3.0
celery==5.3.4

# Async support
asyncio-mqtt==0.16.1
aiofiles==23.2.1
